+++ src/engine/evaluation/imbalance_eval.rs
@@ -97,7 +97,7 @@ mod tests {
     use crate::engine::board::fen::FenTrait;
     use crate::engine::board::structures::color::{BLACK, WHITE};
     use crate::engine::evaluation::init_eval::InitEvalTrait;
-    use crate::engine::evaluation::test_evaluation::SF_EVAL;
+    use crate::engine::evaluation::test_evaluation::{eval_assert, SF_EVAL};
 
     use super::*;
 
@@ -110,7 +110,7 @@ mod tests {
             board.imbalance(WHITE);
             board.imbalance(BLACK);
 
-            assert_eq!(board.calculate_score(), obj.imbalance);
+            eval_assert(board.calculate_score(), obj.imbalance, 0, false);
         }
     }
 }
