+++ src/engine/evaluation/mobility_eval.rs
@@ -106,7 +106,20 @@ impl MobilityEvalTrait for Board {
 mod tests {
 
     use crate::engine::board::fen::FenTrait;
-    use crate::engine::evaluation::test_evaluation::SF_EVAL;
+    use crate::engine::board::structures::color::{BLACK, WHITE};
+    use crate::engine::evaluation::init_eval::InitEvalTrait;
+    use crate::engine::evaluation::test_evaluation::{eval_assert, SF_EVAL};
 
     use super::*;
+
+    #[test]
+    fn mobility_test() {
+        for obj in &SF_EVAL {
+            let mut board = Board::read_fen(obj.fen);
+            board.init();
+            board.mobility_eval(WHITE);
+            board.mobility_eval(BLACK);
+            eval_assert(board.calculate_score(), obj.mobility, 28, false); // FIXME: The difference is too quiet high
+        }
+    }
 }
