+++ src/engine/evaluation/space_eval.rs
@@ -64,7 +64,7 @@ mod tests {
     use crate::engine::board::fen::FenTrait;
     use crate::engine::board::structures::color::{BLACK, WHITE};
     use crate::engine::evaluation::init_eval::InitEvalTrait;
-    use crate::engine::evaluation::test_evaluation::SF_EVAL;
+    use crate::engine::evaluation::test_evaluation::{eval_assert, SF_EVAL};
     use crate::engine::evaluation::trace_eval::TraceEvalTrait;
 
     use super::*;
@@ -72,25 +72,12 @@ mod tests {
     #[test]
     fn space_test() {
         for obj in &SF_EVAL {
-            // 4 and 7
-            if obj.fen != SF_EVAL[4].fen {
-                continue;
-            }
-
             let mut board = Board::read_fen(obj.fen);
             board.init();
             board.space(WHITE);
             board.space(BLACK);
 
-            if board.calculate_score() != obj.space {
-                println!("assertion `{:?} == {:?}` failed", board.calculate_score(), obj.space);
-            } else {
-                println!("assertion `{:?} == {:?}` success", board.calculate_score(), obj.space);
-            }
-            board.print_trace_log("");
-            assert_eq!(board.calculate_score(), obj.space);
-
-            // assert_eq!(board.calculate_score(), obj.space);
+            eval_assert(board.calculate_score(), obj.space, 0, false);
         }
     }
 }
